backend/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
pytz==2022.4
sqlparse==0.4.3
zipp==3.9.0

requests~=2.28.1
djangorestframework-simplejwt==5.2.2
mixer==7.2.2

# drf-yasg:
drf-yasg==1.21.4

# A Django integration for Graphene:
graphene-django==3.0.0

# PostgreSQL database adapter for the Python:
psycopg2==2.9.5

# Gunicorn server for deploy:
gunicorn==20.1.0
